Tidy up and included a method to use different browsers

+++ src/main/java/baseMethods/BaseMethods.java
@@ -6,6 +6,8 @@ import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.edge.EdgeDriver;
+import org.openqa.selenium.firefox.FirefoxDriver;
 import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.ui.ExpectedCondition;
 import org.openqa.selenium.support.ui.ExpectedConditions;
@@ -17,29 +19,80 @@ import org.testng.Assert;
 
 public class BaseMethods {
 
-	public static WebDriver driver;
+	// Write here what browser you want to test with
+			private static String browser = /* write chrome, edge or firefox here ---> */ "edge";
 
-	public static void setDriver() {
-		System.setProperty("webdriver.chrome.driver", "C:\\Users\\JonCo\\git\\Donuts\\Drivers\\chromedriver.exe");
-		driver = new ChromeDriver();
+			// Declare the web driver
+			public static WebDriver driver;
 
-	}
+			// Set the web driver that all scenarios use
+			public static WebDriver setDriver() {
 
-	public static void navigateToHomePage() {
-		driver.manage().window().maximize();
-		// driver.manage().timeouts().pageLoadTimeout(60, TimeUnit.SECONDS);
-		driver.manage().timeouts().pageLoadTimeout(Duration.ofSeconds(60));
-		String homePage = "https://offbeatdonuts.com/";
-		driver.get(homePage);
+				try {
 
-	}
+					String browserName = browser;
 
-	public static void closeDriver() {
-		driver.manage().deleteAllCookies();
-		driver.close();
-		driver.quit();
+					switch (browserName) {
 
-	}
+					// fire fox setup
+					case "firefox":
+						if (null == driver) {
+							System.setProperty("webdriver.gecko.driver",
+									"C:\\Users\\JonCo\\git\\Donuts\\Drivers\\geckodriver.exe");
+							driver = new FirefoxDriver();
+						}
+						break;
+
+					// chrome setup
+					case "chrome":
+						if (null == driver) {
+							System.setProperty("webdriver.chrome.driver",
+									"C:\\Users\\JonCo\\git\\Donuts\\Drivers\\chromedriver.exe");
+							driver = new ChromeDriver();
+						}
+						break;
+
+					// Edge setup
+					case "edge":
+						if (null == driver) {
+
+							System.setProperty("webdriver.edge.driver",
+									"C:\\Users\\JonCo\\git\\Donuts\\Drivers\\msedgedriver.exe");
+							driver = new EdgeDriver();
+						}
+						break;
+					}
+				} catch (Exception e) {
+					System.out.println("Unable to load browser! - Exception: " + e.getMessage());
+				} finally {
+					driver.manage().timeouts().pageLoadTimeout(Duration.ofSeconds(60));
+				}
+				return driver;
+			}
+
+			// open the url
+			public static void navigateToHomePage() {
+				driver.manage().window().maximize();
+				// driver.manage().timeouts().pageLoadTimeout(60, TimeUnit.SECONDS);
+				driver.manage().timeouts().pageLoadTimeout(Duration.ofSeconds(60));
+				String homePage = "https://offbeatdonuts.com/";
+				driver.get(homePage);
+
+			}
+
+			// close the web driver and clear cookies
+			public static void closeDriver() {
+
+				try {
+					if (driver != null) {
+						driver.manage().deleteAllCookies();
+						driver.quit();
+						driver = null;
+					}
+				} catch (Exception e) {
+					System.out.println("Method Failed: screenshotOnFailure, Exception: " + e.getMessage());
+				}
+			}
 
 	//Click an element
 	public static void Click(By locator) {
